Backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
channels==4.0.0
drf-spectacular==0.27.0
django-cors-headers==4.3.1
django-filter==23.5